adk_backend/requirements.txt
# ADK Backend Requirements
google-adk>=0.2.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
httpx>=0.25.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
google-generativeai>=0.3.0
# openai>=1.3.0  # No longer needed - using Google Imagen instead
asyncio-mqtt>=0.16.0
redis>=5.0.0
celery>=5.3.0
prometheus-client>=0.19.0
structlog>=23.2.0
